Update wiki e212 scraper
While it mostly works, there is a bunch of edgecases in the output
that we need to improve handling of, especially with overrides of
MCCs that are shared by multiple countries. It would be nice if the
wiki page was structured a bit differently, as that could make it
possible to extract the country codes from the page directly, but
that seems like a non-trivial task currently.